+++ lib/gg_git_test_helpers.dart
@@ -4,4 +4,5 @@
 // Use of this source code is governed by terms that can be
 // found in the LICENSE file in the root of this package.
 
+export 'src/test_helpers/cached_repos.dart';
 export 'src/test_helpers/test_helpers.dart';